+++ edx_arch_experiments/datadog_monitoring/code_owner/datadog.py
@@ -1,7 +1,7 @@
 """
 Datadog span processor for celery span code owners.
 """
-from .utils import set_code_owner_attribute_from_module
+from .utils import set_code_owner_span_tags_from_module
 
 
 class CeleryCodeOwnerSpanProcessor:
@@ -17,7 +17,7 @@ class CeleryCodeOwnerSpanProcessor:
             # We can use this for celery spans, because the resource name is more predictable
             # and available from the start. For django requests, we'll instead continue to use
             # django middleware for setting code owner.
-            set_code_owner_attribute_from_module(span.resource)
+            set_code_owner_span_tags_from_module(span.resource)
 
     def on_span_finish(self, span):
         pass