+++ edx_arch_experiments/datadog_monitoring/code_owner/utils.py
@@ -1,5 +1,5 @@
 """
-Utilities for monitoring code_owner_2
+Utilities for monitoring code_owner
 """
 import logging
 
@@ -18,9 +18,6 @@ def get_code_owner_from_module(module):
     For example, if the module were 'openedx.features.discounts.views',
     this lookup would match on 'openedx.features.discounts' before
     'openedx.features', because the former is more specific.
-
-    See how to:
-    https://github.com/openedx/edx-django-utils/blob/master/edx_django_utils/monitoring/docs/how_tos/add_code_owner_custom_attribute_to_an_ida.rst
 
     """
     if not module:
@@ -54,7 +51,7 @@ _PATH_TO_CODE_OWNER_MAPPINGS = None
 
 def get_code_owner_mappings():
     """
-    Returns the contents of the CODE_OWNER_MAPPINGS Django Setting, processed
+    Returns the contents of the CODE_OWNER_TO_PATH_MAPPINGS Django Setting, processed
     for efficient lookup by path.
 
     Returns:
@@ -81,12 +78,12 @@ def get_code_owner_mappings():
     # processed map.  Worst case, it is processed more than once at start-up.
     path_to_code_owner_mapping = {}
 
-    # .. setting_name: CODE_OWNER_MAPPINGS
+    # .. setting_name: CODE_OWNER_TO_PATH_MAPPINGS
     # .. setting_default: None
     # .. setting_description: Used for monitoring and reporting of ownership. Use a
     #      dict with keys of code owner name and value as a list of dotted path
     #      module names owned by the code owner.
-    code_owner_mappings = getattr(settings, 'CODE_OWNER_MAPPINGS', None)
+    code_owner_mappings = getattr(settings, 'CODE_OWNER_TO_PATH_MAPPINGS', None)
     if code_owner_mappings is None:
         return None
 
@@ -97,7 +94,7 @@ def get_code_owner_mappings():
                 path_to_code_owner_mapping[path] = code_owner
     except TypeError as e:
         log.exception(
-            'Error processing CODE_OWNER_MAPPINGS. {}'.format(e)  # pylint: disable=logging-format-interpolation
+            'Error processing CODE_OWNER_TO_PATH_MAPPINGS. {}'.format(e)  # pylint: disable=logging-format-interpolation
         )
         raise e
 
@@ -105,54 +102,57 @@ def get_code_owner_mappings():
     return _PATH_TO_CODE_OWNER_MAPPINGS
 
 
-def set_code_owner_attribute_from_module(module):
+def set_code_owner_span_tags_from_module(module):
     """
-    Updates the code_owner_2 and code_owner_2_module custom attributes.
-
-    Celery tasks or other non-web functions do not use middleware, so we need
-        an alternative way to set the code_owner_2 custom attribute.
-
-    Note: These settings will be overridden by the CodeOwnerMonitoringMiddleware.
-        This method can't be used to override web functions at this time.
+    Updates the code_owner and code_owner_module custom span tags.
 
     Usage::
 
-        set_code_owner_2_attribute_from_module(__name__)
+        set_code_owner_span_tags_from_module(__name__)
 
     """
-    set_custom_attribute('code_owner_2_module', module)
+    # .. custom_attribute_name: code_owner_module
+    # .. custom_attribute_description: The module used to determine the code_owner. This can
+    #     be useful for debugging issues for missing code owner span tags.
+    set_custom_attribute('code_owner_module', module)
     code_owner = get_code_owner_from_module(module)
 
     if code_owner:
-        set_code_owner_custom_attributes(code_owner)
+        set_code_owner_custom_span_tags(code_owner)
 
 
-def set_code_owner_custom_attributes(code_owner):
+def set_code_owner_custom_span_tags(code_owner):
     """
-    Sets custom metrics for code_owner_2, code_owner_2_theme, and code_owner_2_squad
+    Sets custom span tags for code_owner, and code_owner_squad
     """
     if not code_owner:  # pragma: no cover
         return
-    set_custom_attribute('code_owner_2', code_owner)
-    theme = _get_theme_from_code_owner(code_owner)
-    if theme:
-        set_custom_attribute('code_owner_2_theme', theme)
-    squad = _get_squad_from_code_owner(code_owner)
-    if squad:
-        set_custom_attribute('code_owner_2_squad', squad)
+
+    # .. custom_attribute_name: code_owner
+    # .. custom_attribute_description: The squad owner name for the tagged span.
+    set_custom_attribute('code_owner', code_owner)
+    # .. custom_attribute_name: code_owner_squad
+    # .. custom_attribute_description: Deprecated code_owner_squad is now redundant
+    #     to the code_owner span tag.
+    set_custom_attribute('code_owner_squad', code_owner)
+
+    # .. custom_attribute_name: code_owner_plugin
+    # .. custom_attribute_description: This is a temporary span tag to roll out the
+    #     the switch from edx-django-utils to this plugin. If this span tag is True,
+    #     the plugin has added the above custom span tags (possibly in addition to
+    #     edx-django-utils). If the code_owner_theme span tag is also seen, then
+    #     edx-django-utils is also adding these span tags. If not, only the plugin is
+    #     creating these tags.
+    set_custom_attribute('code_owner_plugin', True)
 
 
-def set_code_owner_attribute(request):
+def set_code_owner_span_tags_from_request(request):
     """
-    Sets the code_owner_2 custom attribute for the request.
+    Sets the code_owner custom span tag for the request.
     """
-    code_owner = None
     module = _get_module_from_request(request)
     if module:
-        code_owner = get_code_owner_from_module(module)
-
-    if code_owner:
-        set_code_owner_custom_attributes(code_owner)
+        set_code_owner_span_tags_from_module(module)
 
 
 def _get_module_from_request(request):
@@ -160,9 +160,7 @@ def _get_module_from_request(request):
     Get the module from the request path or the current transaction.
 
     Side-effects:
-        Sets code_owner_2_module custom attribute, used to determine code_owner_2.
-        If module was not found, may set code_owner_2_path_error custom attribute
-            if applicable.
+        Sets code_owner_path_error custom span tag if applicable.
 
     Returns:
         str: module name or None if not found
@@ -172,14 +170,14 @@ def _get_module_from_request(request):
         return None
 
     module, path_error = _get_module_from_request_path(request)
-    if module:
-        set_custom_attribute('code_owner_2_module', module)
-        return module
 
-    # monitor errors if module was not found
     if path_error:
-        set_custom_attribute('code_owner_2_path_error', path_error)
-    return None
+        # .. custom_attribute_name: code_owner_path_error
+        # .. custom_attribute_description: Error details if the module can't be found. This can
+        #     be useful for debugging issues for missing code owner span tags.
+        set_custom_attribute('code_owner_path_error', path_error)
+
+    return module
 
 
 def _get_module_from_request_path(request):
@@ -204,100 +202,3 @@ def clear_cached_mappings():
     """
     global _PATH_TO_CODE_OWNER_MAPPINGS
     _PATH_TO_CODE_OWNER_MAPPINGS = None
-    global _CODE_OWNER_TO_THEME_AND_SQUAD_MAPPINGS
-    _CODE_OWNER_TO_THEME_AND_SQUAD_MAPPINGS = None
-
-
-# Cached lookup table for code owner theme and squad given a code owner.
-# - Although code owner is "theme-squad", a hyphen may also be in the theme or squad name, so this ensures we get both
-#   correctly from config.
-# Do not access this directly, but instead use get_code_owner_theme_squad_mappings.
-_CODE_OWNER_TO_THEME_AND_SQUAD_MAPPINGS = None
-
-
-def get_code_owner_theme_squad_mappings():
-    """
-    Returns the contents of the CODE_OWNER_THEMES Django Setting, processed
-    for efficient lookup by path.
-
-    Returns:
-         (dict): dict mapping code owners to a dict containing the squad and theme, or
-            an empty dict if there are no configured mappings.
-
-    Example return value::
-
-        {
-            'theme-x-team-red': {
-                'theme': 'theme-x',
-                'squad': 'team-red',
-            },
-            'theme-x-team-blue': {
-                'theme': 'theme-x',
-                'squad': 'team-blue',
-            },
-        }
-
-    """
-    global _CODE_OWNER_TO_THEME_AND_SQUAD_MAPPINGS
-
-    # Return cached processed mappings if already processed
-    if _CODE_OWNER_TO_THEME_AND_SQUAD_MAPPINGS is not None:
-        return _CODE_OWNER_TO_THEME_AND_SQUAD_MAPPINGS
-
-    # Uses temporary variable to build mappings to avoid multi-threading issue with a partially
-    # processed map.  Worst case, it is processed more than once at start-up.
-    code_owner_to_theme_and_squad_mapping = {}
-
-    # .. setting_name: CODE_OWNER_THEMES
-    # .. setting_default: None
-    # .. setting_description: Used for monitoring and reporting of ownership. Use a
-    #      dict with keys of code owner themes and values as a list of code owner names
-    #      including theme and squad, separated with a hyphen.
-    code_owner_themes = getattr(settings, 'CODE_OWNER_THEMES', {})
-
-    try:
-        for theme in code_owner_themes:
-            code_owner_list = code_owner_themes[theme]
-            for code_owner in code_owner_list:
-                squad = code_owner.split(theme + '-', 1)[1]
-                code_owner_details = {
-                    'theme': theme,
-                    'squad': squad,
-                }
-                code_owner_to_theme_and_squad_mapping[code_owner] = code_owner_details
-    except TypeError as e:
-        log.exception(
-            'Error processing CODE_OWNER_THEMES setting. {}'.format(e)  # pylint: disable=logging-format-interpolation
-        )
-        raise e
-
-    _CODE_OWNER_TO_THEME_AND_SQUAD_MAPPINGS = code_owner_to_theme_and_squad_mapping
-    return _CODE_OWNER_TO_THEME_AND_SQUAD_MAPPINGS
-
-
-def _get_theme_from_code_owner(code_owner):
-    """
-    Returns theme for a code_owner (e.g. 'theme-my-squad' => 'theme')
-    """
-    mappings = get_code_owner_theme_squad_mappings()
-    if mappings is None:  # pragma: no cover
-        return None
-
-    if code_owner in mappings:
-        return mappings[code_owner]['theme']
-
-    return None
-
-
-def _get_squad_from_code_owner(code_owner):
-    """
-    Returns squad for a code_owner (e.g. 'theme-my-squad' => 'my-squad')
-    """
-    mappings = get_code_owner_theme_squad_mappings()
-    if mappings is None:  # pragma: no cover
-        return None
-
-    if code_owner in mappings:
-        return mappings[code_owner]['squad']
-
-    return None